refactor(models): extract Player/Game/WorldMarket codec headroom (#4334)
Move Player JSON + research prune, Game decode, and WorldMarketState
codec/equality into dedicated libraries so near-cap aggregates gain ≥30
lines headroom under the 400 physical ceiling without changing public
types or save JSON shape.

Refs #4334
